% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nino.R
\name{hadsst3}
\alias{hadsst3}
\title{describeIn nino Download SST data from
<https://www.metoffice.gov.uk/hadobs/hadsst3/data/download.html>}
\usage{
hadsst3(index = c("nino12", "nino3", "nino34", "nino4"))
}
\description{
describeIn nino Download SST data from
<https://www.metoffice.gov.uk/hadobs/hadsst3/data/download.html>
}
