% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnino.R
\docType{package}
\name{rnino}
\alias{rnino}
\alias{rnino-package}
\title{rnino: nino indices data package}
\description{
Provides nino indices computed from Met Office Hadley Center HadISST1 and
HadSST3 datasets.
}
\section{Functions}{

\code{\link{hadisst1}}
\code{\link{hadsst3}}
}

