% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nino.R
\name{hadisst1}
\alias{hadisst1}
\title{SST data}
\usage{
hadisst1(index = c("nino12", "nino3", "nino34", "nino4"), anom = FALSE)
}
\arguments{
\item{index}{Character.  Indicates the nino index, possible values are
'nino1+2', 'nino3', 'nino3.4', or 'nino4'.  Default 'nino1+2'.}

\item{anom}{Logical.  Indicates if the index should be anomalies (deviation
from mean) or absolute values.  Only available for HadISST1; default FALSE.}
}
\value{
Dataframe.  Both datasets have a year and month column.  HadISST1
  will have a single value column.  HadSST3 will have 100 value columns, each
  representing a single realization.  A value of -99 indicates a missing
  value.
}
\description{
Download SST data from <https://www.esrl.noaa.gov/psd/gcos_wgsp/Timeseries/>
}
